Only check for authentication in has_permission
This replaces the previous method of checking if the user had an API
key.

This makes no sense, as we only get the user by looking up an API key,
so the check didn't do anything. Also, we have a situation where any
authenticated user has access to all API endpoints, so we only need to
return True if we have a user that's authenticated